refactor: move API routes to separate module
