+++ src/routes/mod.rs
@@ -10,6 +10,7 @@ use tera::{Context, Tera};
 use crate::models::auth::AuthenticatedUser;
 
 pub mod admin;
+pub mod api;
 pub mod auth;
 pub mod main;
